bpo-34272: Move argument parsing tests from test_capi to test_getargs2.
